fix : error핸들링 패키지 정해주기
- 전역이 아니라 사용하는 패키지에 따로 저장
